refactor: 파일위치 수정 및 코드 수정 #32
