+++ src/hooks/weatherHook/useWeatherQueries.ts
@@ -7,12 +7,12 @@ import {
   type CurrentWeatherApiResponse,
   type HourlyWeatherApiResponse,
   type WeeklyWeatherApiResponse,
-} from '../api/weather';
+} from '../../api/weather';
 
-/**
- * 현재 날씨 조회
- * - 백엔드 명세: GET /api/v1/weather/current?locationId={id}
- */
+/*
+현재 날씨 조회
+- 백엔드 명세: GET /api/v1/weather/current?locationId={id}
+*/
 export function useCurrentWeather(locationId?: number) {
   return useQuery<CurrentWeatherApiResponse>({
     queryKey: ['weather', 'current', locationId],
@@ -26,10 +26,11 @@ export function useCurrentWeather(locationId?: number) {
   });
 }
 
-/**
- * 시간대(시간별) 날씨 조회
- * - 백엔드 명세: GET /api/v1/weather/hourly-status?locationId={id}
- */
+/*
+시간대(시간별) 날씨 조회
+- 백엔드 명세: GET /api/v1/weather/hourly?locationId={id}
+- 3시간 간격 데이터를 받아와서 24시간(8개) 분량만 잘라서 반환
+*/
 export function useHourlyWeather(locationId?: number) {
   return useQuery<HourlyWeatherApiResponse>({
     queryKey: ['weather', 'hourly', locationId],
@@ -43,10 +44,10 @@ export function useHourlyWeather(locationId?: number) {
   });
 }
 
-/**
- * 주간 날씨 조회
- * - 백엔드 명세: GET /api/v1/weather/weekly-status?locationId={id}
- */
+/*
+주간 날씨 조회
+- 백엔드 명세: GET /api/v1/weather/weekly?locationId={id}
+*/
 export function useWeeklyWeather(locationId?: number) {
   return useQuery<WeeklyWeatherApiResponse>({
     queryKey: ['weather', 'weekly', locationId],
